Add project settings for default values
Make HideWhenFound and ReadOnlyWhenFound configurable at the project level for the default value when not set
Remove StopSearchWhenFound and make that the way it just works with no configuration. Why are we always searching for the component when we will never use it since we already have one?
Adds a dependency on com.unity.settings-manager